feat: implement task winner evaluation and add scheduled workflow for automatic updates
